Move DisplayExecutor to .ui.util
