+++ plugins/com.google.cloud.tools.eclipse.ui.util.test/src/com/google/cloud/tools/eclipse/ui/util/DisplayExecutorTest.java
@@ -14,10 +14,11 @@
  * the License.
  */
 
-package com.google.cloud.tools.eclipse.dataflow.ui.util;
+package com.google.cloud.tools.eclipse.ui.util;
 
 import static org.mockito.Mockito.verify;
 
+import com.google.cloud.tools.eclipse.ui.util.DisplayExecutor;
 import org.eclipse.swt.widgets.Display;
 import org.junit.Before;
 import org.junit.Test;
